Moving callbacks to VulkanLoadTestSample
