+++ tests/loadtests/vkloadtests/Texture.h
@@ -10,7 +10,6 @@
 
 #include "VulkanLoadTestSample.h"
 
-#include <ktxvulkan.h>
 #include <glm/gtc/matrix_transform.hpp>
 
 class Texture : public VulkanLoadTestSample
